Move boilerplate code to hf_utils.py
